requirements.txt
apache-airflow==2.7.2
requests==2.31.0
